change float to bf16

+++ src/lerobot/policies/fastwam/wan_video_dit.py
@@ -606,7 +606,7 @@ class WanVideoDiT(WanModel):
         }
 
     def post_dit(self, x_tokens: torch.Tensor, pre_state: dict[str, Any]) -> torch.Tensor:
-        x = self.head(x_tokens.to(torch.float32), pre_state["t"].to(torch.float32))
+        x = self.head(x_tokens, pre_state["t"])
         return torch.stack(super().unpatchify(x, pre_state["meta"]["grid_sizes"]))
 
     def forward(
